Move DetermineLicenceEndDate into /jobs/import
