+++ test/services/jobs/import/process-import-licence.service.test.js
@@ -11,7 +11,7 @@ const { expect } = Code
 const { generateUUID } = require('../../../../app/lib/general.lib.js')
 
 // Things we need to stub
-const DetermineLicenceEndDateChangedService = require('../../../../app/services/import/determine-licence-end-date-changed.service')
+const DetermineLicenceEndDateChangedService = require('../../../../app/services/jobs/import/determine-licence-end-date-changed.service.js')
 const ProcessBillingFlagService = require('../../../../app/services/licences/supplementary/process-billing-flag.service')
 const GenerateReturnLogsService = require('../../../../app/services/import/generate-return-logs.service')
